refactor(h707): Improve extension registration and extension docs (#711)
# refactor: Improve extension registration

## Description
This PR improves the `blind-search-tagbox` feature registration with few
important highlights:
- Pure extension flow. Removes mixed registration patterns (manual model
binding, hooks & extension flow) as per the intended extension design
leaving just extension definition with auto-invoked hooks. Result
     - Hook is removed
     - Manual registration in `form-editor` is removed
     - All survey components auto register the extension
- Improves the `add-survey-feature` skill docs to accelerate extensions
delivery and to clear ambiguity
- Removes expensive lazy loaded survey traversal to reduce performance
hits during survey registration
- Simplifies module exports to only what's needed outside of the
extension

## Related Issues
- relates https://github.com/endatix/endatix-hub/issues/707

## Type of Change
Check all that apply.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Documentation update
- [ ] Dependency update
- [x] Refactoring (non-breaking change which improves code quality)
- [ ] Other (please describe)


## Checklist
- My code follows the style guidelines of this project
- I have performed a self-review of my own code
- I have commented my code, particularly in hard-to-understand areas
- I have made corresponding changes to the documentation
- My changes generate no new warnings
- I have added tests that prove my fix is effective or that my feature
works
- New and existing tests pass locally with my changes
- Any dependent changes have been merged and published in downstream
modules

> [!NOTE]
> - [x] I confirm that my Pull Request follows all the checklist
requirements above

## Screenshots
If applicable, add screenshots to help explain your changes.
