+++ lib/survey-features/blind-search-tagbox/index.ts
@@ -1,15 +1,2 @@
-export {
-  BLIND_SEARCH_HANDLERS_ATTACHED_KEY,
-  BLIND_SEARCH_TAGBOX_TYPE,
-  DEFAULT_MIN_SEARCH_LENGTH,
-  EDX_HIDE_UNTIL_TYPING_PROPERTY,
-  EDX_MIN_SEARCH_LENGTH_PROPERTY,
-} from "./constants";
+export { BLIND_SEARCH_TAGBOX_EXTENSION_ID } from "./constants";
 export { blindSearchTagboxExtension } from "./infrastructure/blind-search-tagbox.extension";
-export { registerBlindSearchTagboxGlobals } from "./infrastructure/registry";
-export { bindBlindSearchToSurvey } from "./infrastructure/survey-bindings";
-export { useBlindSearchTagbox } from "./ui/use-blind-search-tagbox.hook";
-export type {
-  BlindSearchRuntimeState,
-  BlindSearchTagboxQuestion,
-} from "./types";
